fix: remove storefront styling from Studio routes
